share tag encoding methods
